+++ packages/core/types.ts
@@ -165,7 +165,6 @@ type RawPlugins = {[id: string]: Plugin}
 type BaseConfig = {
     
     root: string // Root of the project (will resolve config file there)
-    __root?: string // Root configuration file (dynamically set based on builds)
 
     target: TargetType,