+++ packages/core/utils/index.ts
@@ -11,16 +11,22 @@ export const safePath = (path) => path.replace(/^.*?:/, '')
 export const safeJoin = (...paths) => join(paths[0], ...paths.slice(1).map(safePath))
 
 // Get icon safely
-export const getIcon = (icon: ResolvedConfig['icon']) => {
+export const getIcon = (icon: ResolvedConfig['icon'], type?: typeof valid.icon[number] ) => {
     
     if (icon) {
-        let res
-        if (typeof icon === 'string') res = icon
+        if (typeof icon === 'string') return safePath(icon)
         else {
-            const found = valid.icon.find(str => isIconString(icon[str]))
-            res = found ? icon[found] : Object.values(icon).find(isIconString)
-        }
 
-        return res ? safePath(res) : res
+            // Get specified type
+            if (type) {
+                const found = icon[type]
+                if (found) return safePath(found)
+            }
+
+            // Get first valid icon
+            const found = valid.icon.find(str => isIconString(icon[str]))
+            const resolved = found ? icon[found] : Object.values(icon).find(isIconString)
+            return resolved ? safePath(resolved) : resolved
+        }
     } 
 }